+++ test/td_ie/ingests/workflow_test.exs
@@ -2,8 +2,10 @@ defmodule TdIe.Ingests.WorkflowTest do
   use TdIe.DataCase
 
   import Mox
+  import TdIe.TestOperators
 
   alias TdCache.Redix.Stream
+  alias TdCore.Search.MockIndexWorker
   alias TdIe.Ingests
   alias TdIe.Ingests.IngestVersion
   alias TdIe.Ingests.Workflow
@@ -15,8 +17,9 @@ defmodule TdIe.Ingests.WorkflowTest do
 
   setup do
     start_supervised(TdIe.Cache.IngestLoader)
-    start_supervised(TdIe.Search.Cluster)
-    start_supervised(TdIe.Search.IndexWorker)
+    start_supervised!(TdCore.Search.Cluster)
+    start_supervised!(TdCore.Search.IndexWorker)
+
     :ok
   end
 
@@ -30,8 +33,6 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :put_user
 
     test "with valid data creates a ingest", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
-
       domain_id = 1
 
       ingest_attrs = %{
@@ -65,10 +66,10 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert ingest_version.ingest.type == ingest_attrs.type
       assert ingest_version.ingest.domain_id == ingest_attrs.domain_id
       assert ingest_version.ingest.last_change_by == ingest_attrs.last_change_by
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "publishes an audit event including domain_id", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
       domain_id = 1
 
       ingest_attrs = %{
@@ -94,6 +95,7 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, [%{id: ^event_id} = event]} = Stream.read(:redix, @stream, transform: true)
       assert %{payload: payload} = event
       assert %{"ingest" => %{"domain_id" => ^domain_id}} = Jason.decode!(payload)
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "with invalid data returns error changeset" do
@@ -114,7 +116,6 @@ defmodule TdIe.Ingests.WorkflowTest do
     end
 
     test "with content", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
       domain_id = 1
 
       content_schema = [
@@ -147,10 +148,10 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, %{ingest_version: ingest_version}} = Workflow.create_ingest(creation_attrs)
 
       assert %{content: ^content} = ingest_version
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "with invalid content: required", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
       domain_id = 1
 
       content_schema = [
@@ -191,10 +192,10 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert ingest_version.ingest.type == ingest_attrs.type
       assert ingest_version.ingest.domain_id == ingest_attrs.domain_id
       assert ingest_version.ingest.last_change_by == ingest_attrs.last_change_by
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "with content: default values", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
       domain_id = 1
 
       content_schema = [
@@ -227,10 +228,10 @@ defmodule TdIe.Ingests.WorkflowTest do
 
       assert ingest_version.content["Field1"] == "Hello"
       assert ingest_version.content["Field2"] == "World"
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "with invalid content: invalid variable list", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
       domain_id = 1
 
       content_schema = [
@@ -270,6 +271,7 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert ingest_version.ingest.type == ingest_attrs.type
       assert ingest_version.ingest.domain_id == ingest_attrs.domain_id
       assert ingest_version.ingest.last_change_by == ingest_attrs.last_change_by
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "with no content", %{user_id: user_id} do
@@ -365,7 +367,6 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :put_user
 
     test "with valid data updates the ingest_version", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
       ingest_version = insert(:ingest_version)
 
       ingest_attrs = %{
@@ -398,11 +399,10 @@ defmodule TdIe.Ingests.WorkflowTest do
 
       assert object.ingest.id == ingest_version.ingest.id
       assert object.ingest.last_change_by == 1000
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "with valid content data updates the ingest", %{user_id: user_id} do
-      SearchHelpers.expect_bulk_index()
-
       content_schema = [
         %{"name" => "Field1", "type" => "string", "required" => true},
         %{"name" => "Field2", "type" => "string", "required" => true}
@@ -441,6 +441,7 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert %IngestVersion{} = ingest_version
       assert ingest_version.content["Field1"] == "New first field"
       assert ingest_version.content["Field2"] == "Second field"
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "with invalid data returns error changeset" do
@@ -507,10 +508,10 @@ defmodule TdIe.Ingests.WorkflowTest do
     test "creates a new version and sets current to false on previous version", %{
       claims: claims
     } do
-      SearchHelpers.expect_bulk_index()
       ingest_version = insert(:ingest_version, status: "published")
       assert {:ok, res} = Workflow.new_ingest_version(ingest_version, claims)
       assert %{current: %{current: true}, previous: %{current: false}} = res
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "creates a new version and copies the identifier from the previous version one", %{
@@ -518,7 +519,6 @@ defmodule TdIe.Ingests.WorkflowTest do
       template_with_identifier: template_with_identifier,
       identifier_name: identifier_name
     } do
-      SearchHelpers.expect_bulk_index()
       existing_identifier = "00000000-0000-0000-0000-000000000000"
       ingest = build(:ingest, %{type: template_with_identifier.name})
 
@@ -539,13 +539,15 @@ defmodule TdIe.Ingests.WorkflowTest do
                  content: %{^identifier_name => ^existing_identifier}
                }
              } = res
+
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "publishes an event to the audit stream", %{claims: claims} do
-      SearchHelpers.expect_bulk_index()
       ingest_version = insert(:ingest_version, status: "published")
       assert {:ok, %{audit: event_id}} = Workflow.new_ingest_version(ingest_version, claims)
       assert {:ok, [%{id: ^event_id}]} = Stream.read(:redix, @stream, transform: true)
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
   end
 
@@ -553,8 +555,6 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :claims
 
     test "changes the status and audit fields" do
-      SearchHelpers.expect_bulk_index()
-
       %{last_change_at: ts} = ingest_version = insert(:ingest_version, status: "draft")
 
       %{user_id: user_id} = claims = build(:claims, user_id: 987)
@@ -566,21 +566,21 @@ defmodule TdIe.Ingests.WorkflowTest do
                published
 
       assert DateTime.diff(last_change_at, ts, :microsecond) > 0
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "publishes an event to ingests:events", %{claims: claims} do
-      SearchHelpers.expect_bulk_index()
       %{ingest_id: ingest_id, id: id} = ingest_version = insert(:ingest_version)
       assert {:ok, %{event: event_id}} = Workflow.publish_ingest_version(ingest_version, claims)
 
       assert {:ok, [[^event_id, event_data]]} =
                Stream.range(:redix, "ingests:events", event_id, event_id, transform: false)
 
-      assert event_data == ["event", "publish", "id", "#{ingest_id}", "version_id", "#{id}"]
+      assert event_data ||| ["event", "publish", "id", "#{ingest_id}", "version_id", "#{id}"]
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "publishes an event including domain_ids to the audit stream", %{claims: claims} do
-      SearchHelpers.expect_bulk_index()
       ingest_version = insert(:ingest_version, status: "draft")
 
       assert {:ok, %{audit: event_id}} = Workflow.publish_ingest_version(ingest_version, claims)
@@ -588,6 +588,7 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, [%{id: ^event_id} = event]} = Stream.read(:redix, @stream, transform: true)
       assert %{payload: payload} = event
       assert %{"domain_ids" => _domain_ids} = Jason.decode!(payload)
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
   end
 
@@ -595,8 +596,6 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :claims
 
     test "rejects ingest", %{claims: claims} do
-      SearchHelpers.expect_bulk_index()
-
       reason = "Because I want to"
       ingest_version = insert(:ingest_version, status: "pending_approval")
 
@@ -604,10 +603,10 @@ defmodule TdIe.Ingests.WorkflowTest do
                Workflow.reject_ingest_version(ingest_version, reason, claims)
 
       assert %{status: "rejected", reject_reason: ^reason} = ingest_version
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "publishes an event to the audit stream", %{claims: claims} do
-      SearchHelpers.expect_bulk_index()
       reason = "Because I want to"
       ingest_version = insert(:ingest_version, status: "pending_approval")
 
@@ -615,6 +614,7 @@ defmodule TdIe.Ingests.WorkflowTest do
                Workflow.reject_ingest_version(ingest_version, reason, claims)
 
       assert {:ok, [%{id: ^event_id}]} = Stream.read(:redix, @stream, transform: true)
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
   end
 
@@ -622,18 +622,16 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :claims
 
     test "updates the ingest", %{claims: %{user_id: user_id} = claims} do
-      SearchHelpers.expect_bulk_index()
       ingest_version = insert(:ingest_version)
 
       assert {:ok, %{updated: ingest_version}} =
                Workflow.submit_ingest_version(ingest_version, claims)
 
       assert %{status: "pending_approval", last_change_by: ^user_id} = ingest_version
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     test "publishes an event including domain_ids to the audit stream", %{claims: claims} do
-      SearchHelpers.expect_bulk_index()
-
       ingest_version = insert(:ingest_version, status: "draft")
 
       assert {:ok, %{audit: event_id}} = Workflow.submit_ingest_version(ingest_version, claims)
@@ -641,6 +639,7 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, [%{id: ^event_id} = event]} = Stream.read(:redix, @stream, transform: true)
       assert %{payload: payload} = event
       assert %{"domain_ids" => _domain_ids} = Jason.decode!(payload)
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
   end