+++ test/td_ie_web/controllers/ingest_version_controller_test.exs
@@ -4,13 +4,16 @@ defmodule TdIeWeb.IngestVersionControllerTest do
 
   import Mox
 
+  alias TdCore.Search.MockIndexWorker
+
   setup :set_mox_from_context
   setup :verify_on_exit!
 
   setup do
     start_supervised!(TdIe.Cache.IngestLoader)
-    start_supervised!(TdIe.Search.Cluster)
-    start_supervised!(TdIe.Search.IndexWorker)
+    start_supervised!(TdCore.Search.Cluster)
+    start_supervised!(TdCore.Search.IndexWorker)
+
     :ok
   end
 
@@ -172,7 +175,6 @@ defmodule TdIeWeb.IngestVersionControllerTest do
   describe "create ingest" do
     @tag authentication: [role: "admin"]
     test "renders ingest when data is valid", %{conn: conn, swagger_schema: schema} do
-      SearchHelpers.expect_bulk_index()
       %{id: domain_id, name: domain_name} = CacheHelpers.put_domain()
 
       creation_attrs = %{
@@ -206,6 +208,7 @@ defmodule TdIeWeb.IngestVersionControllerTest do
 
       assert data["domain"]["id"] == domain_id
       assert data["domain"]["name"] == domain_name
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
 
     @tag authentication: [role: "admin"]
@@ -303,8 +306,6 @@ defmodule TdIeWeb.IngestVersionControllerTest do
   describe "create new versions" do
     @tag authentication: [role: "admin"]
     test "create new version with modified template", %{conn: conn} do
-      SearchHelpers.expect_bulk_index()
-
       template_content = [
         %{
           "name" => "group",
@@ -345,14 +346,14 @@ defmodule TdIeWeb.IngestVersionControllerTest do
                  Routes.ingest_version_ingest_version_path(conn, :version, ingest_version.id)
                )
                |> json_response(:created)
+
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
   end
 
   describe "update ingest_version" do
     @tag authentication: [role: "admin"]
     test "renders ingest_version when data is valid", %{conn: conn, swagger_schema: schema} do
-      SearchHelpers.expect_bulk_index()
-
       %{user_id: user_id} = build(:claims)
       %{id: id} = insert(:ingest_version, last_change_by: user_id)
 
@@ -379,6 +380,7 @@ defmodule TdIeWeb.IngestVersionControllerTest do
                |> json_response(:ok)
 
       Enum.each(update_attrs, &assert(Map.get(data, elem(&1, 0)) == elem(&1, 1)))
+      assert [{:reindex, :ingests, [_]}] = MockIndexWorker.calls()
     end
   end